Refactor file namings and route naming for presigned url
